pyproject.toml
[project]
name = "kronos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas==2.2.2",
    "torch==2.6.0",
    "einops==0.8.1",
    "huggingface_hub==0.33.1",
    "matplotlib==3.9.3",
    "tqdm==4.67.1",
    "safetensors>=0.4.0",
    "ccxt>=4.5.1",
    "mplfinance>=0.12.10b0",
    "tushare>=1.4.24",
    "python-dotenv>=1.1.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
